+++ src/outputs.rs
@@ -2,6 +2,7 @@ use iced::{
     Task,
     platform_specific::shell::commands::layer_surface::{
         Anchor, KeyboardInteractivity, Layer, destroy_layer_surface, get_layer_surface, set_anchor,
+        set_exclusive_zone, set_size,
     },
     runtime::platform_specific::wayland::layer_surface::{IcedOutput, SctkLayerSurfaceSettings},
     window::Id,
@@ -11,7 +12,7 @@ use wayland_client::protocol::wl_output::WlOutput;
 
 use crate::{
     HEIGHT,
-    config::{self, Position},
+    config::{self, AppearanceStyle, Position},
     menu::{Menu, MenuType},
     position_button::ButtonUIRef,
 };
@@ -20,6 +21,7 @@ use crate::{
 struct ShellInfo {
     id: Id,
     position: Position,
+    style: AppearanceStyle,
     menu: Menu,
 }
 
@@ -32,8 +34,11 @@ pub enum HasOutput<'a> {
 }
 
 impl Outputs {
-    pub fn new<Message: 'static>(position: Position) -> (Self, Task<Message>) {
-        let (id, menu_id, task) = Self::create_output_layers(None, position);
+    pub fn new<Message: 'static>(
+        style: AppearanceStyle,
+        position: Position,
+    ) -> (Self, Task<Message>) {
+        let (id, menu_id, task) = Self::create_output_layers(style, None, position);
 
         (
             Self(vec![(
@@ -42,6 +47,7 @@ impl Outputs {
                     id,
                     menu: Menu::new(menu_id),
                     position,
+                    style,
                 }),
                 None,
             )]),
@@ -49,18 +55,30 @@ impl Outputs {
         )
     }
 
+    fn get_height(style: AppearanceStyle) -> u32 {
+        HEIGHT
+            - match style {
+                AppearanceStyle::Solid | AppearanceStyle::Gradient => 8,
+                AppearanceStyle::Islands => 0,
+            }
+    }
+
     fn create_output_layers<Message: 'static>(
+        style: AppearanceStyle,
         wl_output: Option<WlOutput>,
         position: Position,
     ) -> (Id, Id, Task<Message>) {
         let id = Id::unique();
+        let height = Self::get_height(style);
+
         let task = get_layer_surface(SctkLayerSurfaceSettings {
             id,
-            size: Some((None, Some(HEIGHT))),
+            namespace: "ashell-main-layer".to_string(),
+            size: Some((None, Some(height))),
             layer: Layer::Bottom,
             pointer_interactivity: true,
             keyboard_interactivity: KeyboardInteractivity::None,
-            exclusive_zone: HEIGHT as i32,
+            exclusive_zone: height as i32,
             output: wl_output.clone().map_or(IcedOutput::Active, |wl_output| {
                 IcedOutput::Output(wl_output)
             }),
@@ -75,6 +93,7 @@ impl Outputs {
         let menu_id = Id::unique();
         let menu_task = get_layer_surface(SctkLayerSurfaceSettings {
             id: menu_id,
+            namespace: "ashell-main-layer".to_string(),
             size: Some((None, None)),
             layer: Layer::Background,
             pointer_interactivity: true,
@@ -137,6 +156,7 @@ impl Outputs {
 
     pub fn add<Message: 'static>(
         &mut self,
+        style: AppearanceStyle,
         request_outputs: &config::Outputs,
         position: Position,
         name: &str,
@@ -147,7 +167,8 @@ impl Outputs {
         if target {
             debug!("Found target output, creating a new layer surface");
 
-            let (id, menu_id, task) = Self::create_output_layers(Some(wl_output.clone()), position);
+            let (id, menu_id, task) =
+                Self::create_output_layers(style, Some(wl_output.clone()), position);
 
             let destroy_task = match self
                 .0
@@ -176,6 +197,7 @@ impl Outputs {
                     id,
                     menu: Menu::new(menu_id),
                     position,
+                    style,
                 }),
                 Some(wl_output),
             ));
@@ -212,6 +234,7 @@ impl Outputs {
 
     pub fn remove<Message: 'static>(
         &mut self,
+        style: AppearanceStyle,
         position: Position,
         wl_output: WlOutput,
     ) -> Task<Message> {
@@ -240,7 +263,7 @@ impl Outputs {
                 if !self.0.iter().any(|(_, shell_info, _)| shell_info.is_some()) {
                     debug!("No outputs left, creating a fallback layer surface");
 
-                    let (id, menu_id, task) = Self::create_output_layers(None, position);
+                    let (id, menu_id, task) = Self::create_output_layers(style, None, position);
 
                     self.0.push((
                         None,
@@ -248,6 +271,7 @@ impl Outputs {
                             id,
                             menu: Menu::new(menu_id),
                             position,
+                            style,
                         }),
                         None,
                     ));
@@ -263,6 +287,7 @@ impl Outputs {
 
     pub fn sync<Message: 'static>(
         &mut self,
+        style: AppearanceStyle,
         request_outputs: &config::Outputs,
         position: Position,
     ) -> Task<Message> {
@@ -303,16 +328,23 @@ impl Outputs {
         debug!("Adding outputs: {:?}", to_add);
 
         let mut tasks = Vec::new();
+
         for (name, wl_output) in to_add {
             if let Some(wl_output) = wl_output {
                 if let Some(name) = name {
-                    tasks.push(self.add(request_outputs, position, name.as_str(), wl_output));
+                    tasks.push(self.add(
+                        style,
+                        request_outputs,
+                        position,
+                        name.as_str(),
+                        wl_output,
+                    ));
                 }
             }
         }
 
         for wl_output in to_remove {
-            tasks.push(self.remove(position, wl_output));
+            tasks.push(self.remove(style, position, wl_output));
         }
 
         for shell_info in self.0.iter_mut().filter_map(|(_, shell_info, _)| {
@@ -341,7 +373,39 @@ impl Outputs {
             ));
         }
 
+        for shell_info in self.0.iter_mut().filter_map(|(_, shell_info, _)| {
+            if let Some(shell_info) = shell_info {
+                if shell_info.style != style {
+                    Some(shell_info)
+                } else {
+                    None
+                }
+            } else {
+                None
+            }
+        }) {
+            debug!(
+                "Change style for output: {:?}, new style {:?}",
+                shell_info.id, style
+            );
+            shell_info.style = style;
+            let height = Self::get_height(style);
+            tasks.push(Task::batch(vec![
+                set_size(shell_info.id, None, Some(height)),
+                set_exclusive_zone(shell_info.id, height as i32),
+            ]));
+        }
+
         Task::batch(tasks)
+    }
+
+    pub fn menu_is_open(&self) -> bool {
+        self.0.iter().any(|(_, shell_info, _)| {
+            shell_info
+                .as_ref()
+                .map(|shell_info| shell_info.menu.menu_info.is_some())
+                .unwrap_or_default()
+        })
     }
 
     pub fn toggle_menu<Message: 'static>(